:recycle: restructure read zx code
